dpo: a validator against a running browser
During development of DPO templates (and the engine itself),
it is useful to repeatedly run the DPO matching part against
a browser that wont re-load the target page. Thus, use some
'behave' part to keep the page open (according to the site),
and then repeatedly attach another process validating DPOs.