pagelems: introduce 'DOM context' while iterating components
Component objects need to reference a key parent object, such
as the 'html' boundary. This will grant them access to local
variables, like the gallery of templates, slot assignments etc.

This has to follow the lifetime of components, not the runtime
call stack; thus an extra argument to the iteration is the best
possible implementation.