pagelems: 'xpath_locator()' API to attempt a specific xpath
Replacing the 'must_have()' API, children elements are set
to be mandatory; but what is the correct xpath to locate a
pagelem? Specifying the element itself may be too broad;
traversing the full children tree may be too much and slow.
Thus, this API assigns a "score" to how specific an xpath
locator may be. Not an exact science, but an 'id=' attribute
is expected to be very specific, the Anyelement '*' not.