pagelems: solution for text+element nodes in DOM
Given the following (remote) DOM:
   <div class="field">
	Field Name: <span>value</span>
   </div>

we should be able to describe in in this very way, either when matching
the text "Field Name:" or reading it.

(<div>).text => 'Field Name: value' , doesn't work
Then, when matching try a partial (contains()) match.
When reading, JS needs to be used to traverse direct children of this
node, extracting text only from TEXT_NODEs and stopping at known elem.