dom_components: not picklable
At base_parsers.py:424, the parent component of a WebDriverException
is stored in it, to aid screenshots. But `behave_parallel` wants to
pickle that Exception when returning step status.
A DOM component cannot be serialized (pickled), so far, because it
relies heavily on the external WebElement, along with the complete
state of the pagelems+scopes. Impossible to carry over, then.