pagelems: descriptors for attributes API
Rather than custom getters/setters, use the descriptor[1] API for
component attributes.
Technically speaking, descriptors *cannot* be dropped into the
`ComponentProxy` classes, because they are designed for the class,
not the instance. But this can be worked around.

Then, the API for attaching attributes to a component becomes more
compact, and also allows Scopes to set writable/deletable attributes,
too.

[1] https://docs.python.org/3/howto/descriptor.html